requirements.txt
python-dotenv
openai
pandas
scikit-learn
python-dotenv
streamlit
groq
sentence-transformers

